tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/multiple_matching_keyframes.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - multiple @keyframes rules with the same name</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#keyframes">
   <meta name="flags" content="dom">
   <meta name="assert" content="The @keyframes rule that is used by an animation
   will be the last one encountered in sorted rules order that matches the name
   of the animation specified by the 'animation-name' property">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes slide {
         0% {
            left: 100px;
         }
         100% {
            left: 0px;
         }
      }
      @keyframes slide {
         0% {
            left: 0px;
         }
         100% {
            left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .test {
         animation: slide 1s forwards;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var passed;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationend', function() {
            clearTimeout(timeout);
            var passed = (getComputedStyle(div, null).left == '100px');
            test(function(){assert_true(passed);}, testName); done();
            log.data = passed ? 'PASS if the animation occurred.' : 'FAIL';
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The square below should slide right for one second.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
